Move jmh generators to testFixtures
As part of another suite, it'd be handy to reference the keys generated here without duplicating.

This change will allow me to import that.
