--- eth-benchmark-tests/src/jmh/java/tech/pegasys/teku/benchmarks/gen/Utils.java
@@ -1,43 +1,0 @@
-/*
- * Copyright Consensys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.benchmarks.gen;
-
-import java.util.Iterator;
-import java.util.NoSuchElementException;
-import java.util.function.Supplier;
-
-public class Utils {
-
-  static <T> Iterator<T> fromSupplier(Supplier<T> supplier) {
-    return new Iterator<T>() {
-
-      T next = supplier.get();
-
-      @Override
-      public boolean hasNext() {
-        return next != null;
-      }
-
-      @Override
-      public T next() {
-        if (next == null) {
-          throw new NoSuchElementException();
-        }
-        T ret = next;
-        next = supplier.get();
-        return ret;
-      }
-    };
-  }
-}
